+++ DB/DBZ/standard/business/src/test/java/jp/co/ndensan/reams/db/dbz/business/core/basic/NinteiChosaScheduleMemoIdentifierTest.java
@@ -27,7 +27,7 @@ public class NinteiChosaScheduleMemoIdentifierTest extends DbzTestBase {
 //TODO 主キーの数が足りない場合、追加してください。
     private static FlexibleDate 主キー名1;
     private static Code 主キー名2;
-    private static Decimal 主キー名3;
+    private static int 主キー名3;
 
     @BeforeClass
     public static void setUpClass() {
